% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ccdf_labels.R
\docType{data}
\name{ccdf_labels}
\alias{ccdf_labels}
\title{Child Care and Development Fund labels data}
\format{Data frame with columns
\describe{
\item{long, lat}{Longitude and latitude}
\item{state_name, state_abbv}{State name and state abbreviation}
}}
\usage{
ccdf_labels
}
\description{
Child Care and Development Fund labels data
}
\keyword{datasets}
