% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ccdfmap.R
\docType{data}
\name{ccdfmap}
\alias{ccdfmap}
\title{Child Care and Development Fund map data}
\format{Data frame with columns
\describe{
\item{long, lat}{Longitude and latitude}
\item{state_name, state_abbv, state_fips}{State name, state abbreviation, two-digit FIPS code}
}}
\usage{
ccdfmap
}
\description{
Child Care and Development Fund map data
}
\keyword{datasets}
