% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/counties.R
\docType{data}
\name{counties}
\alias{counties}
\title{County shapefile data}
\format{Data frame with columns
\describe{
\item{long, lat}{Longitude and latitude}
\item{county_name, county_fips, fips_class}{County name, five-digit FIPS and FIPS class code}
\item{state_name, state_abbv, state_fips}{State name, postal abbreviation, and two-digit FIPS code}
}}
\source{
United States Census Bureau,
 \url{https://www.census.gov/geo/maps-data/data/tiger-cart-boundary.html}
}
\usage{
counties
}
\description{
County shapefile data, cleaned and parsed into a tibble for easy mapping.
Includes various state identifiers for easy merging.
}
\keyword{datasets}
