% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/counties_labels.R
\docType{data}
\name{counties_labels}
\alias{counties_labels}
\title{County labels data}
\format{Data frame with columns
\describe{
\item{long, lat}{Longitude and latitude}
\item{county_name, county_fips, fips_class}{County name, five-digit FIPS and FIPS class code}
\item{state_name, state_abbv, state_fips}{State name, postal abbreviation, and two-digit FIPS code}
}}
\usage{
counties_labels
}
\description{
Includes various state identifiers for easy merging.
}
\keyword{datasets}
