% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/countydata.R
\docType{data}
\name{countydata}
\alias{countydata}
\title{County data for mapping}
\format{Data frame with columns
\describe{
\item{year}{Year of data}
\item{county_fips}{Five-digit FIPS code}
\item{hhhpop}{Household population}
\item{horate}{Homeownership rate}
\item{medhhincome}{Median household income}
}}
\source{
Urban Institute Sloan ADRF Database,
 \url{https://adrf.urban.org}
}
\usage{
countydata
}
\description{
County-level data to use for example maps.
Includes county FIPS code to match with county shapefile data.
}
\keyword{datasets}
