% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_shapefile.R
\name{get_shapefile}
\alias{get_shapefile}
\title{Obtain shapefile from census.gov}
\format{Data frame with columns}
\source{
United States Census Bureau,
 \url{https://www.census.gov/geo/maps-data/data/tiger-cart-boundary.html}
}
\usage{
get_shapefile(year, fips, level)
}
\arguments{
\item{year}{shapefile year}

\item{fips}{state fips}

\item{level}{shapefile level}
}
\description{
Obtain shapefile from census.gov
}
