% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statedata.R
\docType{data}
\name{statedata}
\alias{statedata}
\title{State data for mapping}
\format{Data frame with columns
\describe{
\item{year}{Year of data}
\item{state_fips, state_name}{Two-digit FIPS code and state name}
\item{hhhpop}{Household population}
\item{horate}{Homeownership rate}
\item{medhhincome}{Median household income}
}}
\source{
Urban Institute Sloan ADRF Database,
 \url{https://adrf.urban.org}
}
\usage{
statedata
}
\description{
State-level data to use for example maps.
Includes state FIPS code to match with state shapefile data.
}
\keyword{datasets}
