% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/states.R
\docType{data}
\name{states}
\alias{states}
\title{State shapefile data}
\format{Data frame with columns
\describe{
\item{long,lat}{Longitude and latitude}
\item{state_name,state_abbv,state_fips}{State name, postal abbreviation, and two-digit FIPS code}
}}
\source{
United States Census Bureau,
 \url{https://www.census.gov/geo/maps-data/data/tiger-cart-boundary.html}
}
\usage{
states
}
\description{
State shapefile data, cleaned and parsed into a tibble for easy mapping.
Includes various state identifiers for easy merging.
}
\keyword{datasets}
