% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/states_labels.R
\docType{data}
\name{states_labels}
\alias{states_labels}
\title{States labels data}
\format{Data frame with columns
\describe{
\item{long, lat}{Longitude and latitude}
\item{state_name, state_abbv}{State name and state abbreviation}
}}
\usage{
states_labels
}
\description{
States labels data
}
\keyword{datasets}
