% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/territories_labels.R
\docType{data}
\name{territories_labels}
\alias{territories_labels}
\title{Territories labels data}
\format{Data frame with columns
\describe{
\item{long, lat}{Longitude and latitude}
\item{state_name, state_abbv}{State name and state abbreviation}
}}
\usage{
territories_labels
}
\description{
Territories labels data
}
\keyword{datasets}
