% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_urbn_labels.R
\name{get_urbn_labels}
\alias{get_urbn_labels}
\title{Import different maps}
\usage{
get_urbn_labels(map)
}
\arguments{
\item{map}{Selection of custom labels. Current option is \code{"ccdf"}.}
}
\description{
\code{get_urbn_labels()} loads labels and coordinates for maps from the \code{get_urbn_map()} functions.
}
