% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_urbn_map.R
\name{get_urbn_map}
\alias{get_urbn_map}
\title{Import different maps}
\usage{
get_urbn_map(map = "states")
}
\arguments{
\item{map}{Selection of custom map. Current options are \code{"states"}, \code{"counties"}, \code{"ccdf"}, \code{"territories"}, and \code{"territories_counties"}.}
}
\description{
\code{get_urbn_map()} loads maps that are different than \code{states} and \code{counties}
}
